Increase tic_t from 32-bit to 64-bit; Add wrap around handling for I_GetTimeMS() so when wrap around occurs it is corrected;

+++ src/doomtype.h
@@ -427,7 +427,7 @@ BPLWRITE_x(UInt64,uint64_t)
 ************/
 
 // A valid tic
-typedef uint32_t tic_t;
+typedef uint64_t tic_t;
 
 /******************************************
 *** REMOVE ALL THIS GARBAGE, SERIOUSLY! ***
